move proof to its own package
